fix: convert file.js to file.ts and remove unused functions
