+++ src/utils/spec/file.spec.ts
@@ -5,8 +5,6 @@ import {
   fileExists,
   deleteFile,
   deleteFolder,
-  unifyFilePath,
-  getRelativePath,
   generateTimestamp
 } from '@sasjs/utils'
 
@@ -41,44 +39,6 @@ describe('file utility', () => {
       await expect(readFile(filePath)).resolves.toEqual(content)
 
       deleteFolder(path.join(process.cwd(), 'testFolder_1'))
-    })
-  })
-
-  describe('unifyFilePath', () => {
-    it('should unify file path in Unix-like systems', () => {
-      const filePath = '/folder/subFolder/file.txt'
-
-      expect(unifyFilePath(filePath, '/')).toEqual(filePath)
-    })
-
-    it('should unify file path in Windows system', () => {
-      const filePath = '/folder/subFolder/file.txt'
-      const expectedFilePath = filePath.replace(/\//g, '\\')
-
-      expect(unifyFilePath(filePath, '\\')).toEqual(expectedFilePath)
-    })
-
-    it('should unify file path with mixed separators', () => {
-      const filePath = '/folder/subFolder\\file.txt'
-      const unixSeparator = '/'
-      const winSeparator = '\\'
-      const expectedUnixPath = filePath.replace(/\\/g, unixSeparator)
-      const expectedWinPath = filePath.replace(/\//g, winSeparator)
-
-      expect(unifyFilePath(filePath, unixSeparator)).toEqual(expectedUnixPath)
-      expect(unifyFilePath(filePath, winSeparator)).toEqual(expectedWinPath)
-    })
-
-    it('should return file path with custom separator', () => {
-      let filePath = '/folder/subFolder/file.txt'
-      let expectedFilePath = filePath.replace(/\//g, '$')
-
-      expect(unifyFilePath(filePath, '$')).toEqual(expectedFilePath)
-
-      filePath = '\\folder\\subFolder\\file.txt'
-      expectedFilePath = filePath.replace(/\\/g, '$')
-
-      expect(unifyFilePath(filePath, '$', '\\')).toEqual(expectedFilePath)
     })
   })
 })
